requirements.txt
ollama
psycopg2-binary
sqlalchemy
pgvector
langchain
javalang
tqdm
